validation and NodeSetEntry moved to their own classes
